+++ src/ConvNetSharp.Flow/Ops/Mult.cs
@@ -1,18 +1,20 @@
 using System;
+using System.Linq;
 using ConvNetSharp.Volume;
 
 namespace ConvNetSharp.Flow.Ops
 {
     /// <summary>
-    ///     y = a + b
+    ///     Element wise multiplication
+    ///     y = left * right
     /// </summary>
     /// <typeparam name="T"></typeparam>
-    public class AddOp<T> : Op<T> where T : struct, IEquatable<T>, IFormattable
+    public class Mult<T> : Op<T> where T : struct, IEquatable<T>, IFormattable
     {
         private readonly Op<T> _left;
         private readonly Op<T> _right;
 
-        public AddOp(Op<T> left, Op<T> right)
+        public Mult(Op<T> left, Op<T> right)
         {
             this._left = left;
             this._right = right;
@@ -21,12 +23,12 @@ namespace ConvNetSharp.Flow.Ops
             AddParent(right);
         }
 
-        public override string Representation => "+";
+        public override string Representation => "*";
 
         public override void Differentiate()
         {
-            this._left.RegisterDerivate(this.Derivate);
-            this._right.RegisterDerivate(this.Derivate);
+            this._left.RegisterDerivate(this.Derivate * this._right);
+            this._right.RegisterDerivate(this.Derivate * this._left);
         }
 
         protected override void Dispose(bool disposing)
@@ -58,14 +60,20 @@ namespace ConvNetSharp.Flow.Ops
                 this.Result = BuilderInstance<T>.Volume.SameAs(shape);
             }
 
-            left.DoAdd(right, this.Result);
+            left.DoMultiply(right, this.Result);
 
             return this.Result;
         }
 
         public override string ToString()
         {
-            return $"{this._left} + {this._right}";
+            var addParenthesis = this._left.Parents.Any();
+            var leftStr = addParenthesis ? $"({this._left})" : $"{this._left}";
+
+            addParenthesis = this._right.Parents.Any();
+            var rightStr = addParenthesis ? $"({this._right})" : $"{this._right}";
+
+            return $"{leftStr} * {rightStr}";
         }
     }
 }
